Added a Vector struct to HPCG.  This struct is useful for making accelerator version of the code easier to implement. Added files for the multigrid option.  Code is not working at this point.
M    src/TestSymmetry.hpp
M    src/ComputeSYMGS_ref.hpp
M    src/ComputeWAXPBY_ref.cpp
M    src/ComputeWAXPBY_ref.hpp
M    src/GenerateProblem.cpp
M    src/ComputeSPMV_ref.cpp
M    src/GenerateProblem.hpp
M    src/OptimizeProblem.cpp
M    src/ComputeSPMV_ref.hpp
M    src/ComputeDotProduct.cpp
M    src/OptimizeProblem.hpp
M    src/ExchangeHalo.cpp
M    src/CGData.hpp
M    src/CG.cpp
M    src/ComputeDotProduct.hpp
M    src/TestCG.cpp
M    src/ExchangeHalo.hpp
M    src/CG.hpp
M    src/TestCG.hpp
M    src/ComputeSYMGS.cpp
M    src/Geometry.hpp
A    src/ComputeMG.cpp
A    src/MGData.hpp
M    src/ComputeSYMGS.hpp
M    src/ComputeResidual.cpp
A    src/GenerateGridTransfer.cpp
A    src/ComputeMG.hpp
M    src/ComputeWAXPBY.cpp
A    src/Vector.hpp
M    src/ComputeResidual.hpp
A    src/GenerateGridTransfer.hpp
M    src/ComputeWAXPBY.hpp
M    src/ComputeDotProduct_ref.cpp
M    src/ComputeSPMV.cpp
M    src/CG_ref.cpp
M    src/ComputeDotProduct_ref.hpp
M    src/ComputeSPMV.hpp
M    src/TestSymmetry.cpp
M    src/CG_ref.hpp
M    src/ComputeSYMGS_ref.cpp
M    testing/main.cpp

git-svn-id: svn+ssh://software.sandia.gov/svn/hpcg/trunk@208 5e06a57e-6b39-4354-92a0-249a7d0c5410

+++ src/Geometry.hpp
@@ -34,8 +34,8 @@ typedef int local_int_t;
 
   Define as "long long" when global problem dimension is > 2^31
 */
-//typedef int global_int_t;
-typedef long long global_int_t;
+typedef int global_int_t;
+//typedef long long global_int_t;
 
 // This macro should be defined if the global_int_t is not long long
 // in order to stop complaints from non-C++11 compliant compilers.
